+++ src/test/ui/const-generics/defaults/complex-unord-param.rs
@@ -5,14 +5,6 @@
 #![allow(incomplete_features)]
 #![allow(dead_code)]
 
-struct FixedOutput<'a, const N: usize, T=u32> {
-  out: &'a [T; N],
-}
-
-trait FixedOutputter {
-  fn out(&self) -> FixedOutput<'_, 10>;
-}
-
 struct NestedArrays<'a, const N: usize, A: 'a, const M: usize, T:'a =u32> {
   args: &'a [&'a [T; M]; N],
   specifier: A,
