+++ src/test/ui/const-generics/defaults/needs-feature.none.stderr
@@ -1,17 +1,17 @@
 error: type parameters must be declared prior to const parameters
-  --> $DIR/needs-feature.rs:4:26
+  --> $DIR/needs-feature.rs:10:26
    |
 LL | struct A<const N: usize, T=u32>(T);
    |         -----------------^----- help: reorder the parameters: lifetimes, then types: `<T, const N: usize>`
 
 error[E0658]: const generics are unstable
-  --> $DIR/needs-feature.rs:4:16
+  --> $DIR/needs-feature.rs:10:16
    |
 LL | struct A<const N: usize, T=u32>(T);
    |                ^
    |
-   = note: see issue #44580 <https://github.com/rust-lang/rust/issues/44580> for more information
-   = help: add `#![feature(const_generics)]` to the crate attributes to enable
+   = note: see issue #74878 <https://github.com/rust-lang/rust/issues/74878> for more information
+   = help: add `#![feature(min_const_generics)]` to the crate attributes to enable
 
 error: aborting due to 2 previous errors
