--- src/test/ui/const-generics/defaults/right-order.rs
@@ -1,11 +1,0 @@
-// run-pass
-// Verifies that having generic parameters after constants is permitted.
-
-#![feature(const_generics)]
-#![allow(incomplete_features)]
-
-struct A<const N: usize, T=u32>(T);
-
-fn main() {
-  let _: A<3> = A(0);
-}
