+++ src/test/ui/const-generics/type-after-const-ok.rs
